+++ benchmarks/vote/clients/mod.rs
@@ -13,8 +13,8 @@ pub(crate) trait VoteClient {
 
     fn new(&Parameters, &clap::ArgMatches) -> Self::Constructor;
     fn from(&mut Self::Constructor) -> Self;
-    fn handle_reads(&mut self, requests: &[(time::Instant, usize)]);
-    fn handle_writes(&mut self, requests: &[(time::Instant, usize)]);
+    fn handle_reads(&mut self, requests: &[(time::Instant, i32)]);
+    fn handle_writes(&mut self, requests: &[(time::Instant, i32)]);
 
     fn spawns_threads() -> bool {
         false
